+++ bin/meshroom_photogrammetry
@@ -61,5 +61,5 @@ toNodes = None
 if args.toNode:
     toNodes = graph.findNodes(args.toNode)
 
-meshroom.core.graph.execute(graph, toNodes=toNodes, forceCompute=args.forceCompute, forceStatus=args.forceStatus)
+meshroom.core.graph.executeGraph(graph, toNodes=toNodes, forceCompute=args.forceCompute, forceStatus=args.forceStatus)
